requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-socketio==5.9.1
python-engineio==4.4.0
aiofiles==23.1.0
pandas==2.2.2
python-multipart==0.0.6
jinja2==3.1.2
